Replace aws_account's service_account_id with role_external_id (#48)
CloudSecure's CloudFormation currently uses the CloudSecure service
account's ID as the `sts:ExternalId` in the IAM role definition to add
another layer of security. It would be more secure to use another
secret. Replace the `service_account_id` attribute with
`role_external_id` to support passing in that secret.

Passing the `service_account_id` is not required because it is already
included in the OAuth 2 access token.